+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * pool: add `RelayPool::start` ([Yuki Kishimoto])
 * sdk: add `Client::builder()` ([Yuki Kishimoto])
 * sdk: add `Client::update_min_pow_difficulty` method ([Yuki Kishimoto])
+* ffi(nostr): add `gift_wrap_from_seal` func ([Yuki Kishimoto])
 
 ### Fixed